feat: Convert react-solid-state to typescript
