+++ packages/react-solid-state/babel.config.js
@@ -1,3 +1,7 @@
 module.exports = {
-  presets: [['@babel/preset-env', {targets: {'node': 'current'}}], '@babel/preset-react'],
+  presets: [
+    ['@babel/preset-env', {targets: {'node': 'current'}}],
+    '@babel/preset-react',
+    '@babel/preset-typescript',
+  ],
 };
