+++ packages/react-solid-state/rollup.config.js
@@ -1,7 +1,8 @@
 import nodeResolve from '@rollup/plugin-node-resolve';
+import babel from '@rollup/plugin-babel';
 
 export default {
-  input: 'src/index.js',
+  input: 'src/index.ts',
   output: {
     file: 'lib/react-solid-state.js',
     format: 'cjs',
@@ -9,6 +10,12 @@ export default {
   },
   external: ['react', 'react-dom', 'solid-js'],
   plugins: [
-    nodeResolve({ extensions: ['.js'] })
+    nodeResolve({ extensions: ['.js', '.ts'] }),
+    babel({
+      extensions: ['.js', '.ts'],
+      babelHelpers: "bundled",
+      presets: ["@babel/preset-typescript"],
+      exclude: 'node_modules/**'
+    }),
   ]
 };
